+++ packages/plugin-testing-integration/src/configuration.js
@@ -18,6 +18,8 @@ let configuration = {
     script.text &&
     script.id !== 'ima-runner' &&
     window.eval(script.text),
+  processEnvironment: env => env,
+  applicationFolder: undefined,
 };
 
 /**