+++ packages/cli-plugin-scramble-css/CHANGELOG.md
@@ -1,5 +1,27 @@
 # Change Log
 
+## 1.2.5
+
+### Patch Changes
+
+- 110c2af: Removed invalid RC version ranges in package.json files
+
+## 1.2.4
+
+### Patch Changes
+
+- dc22ea6: Fixed invalid peer dependency
+
+## 1.2.3
+
+### Patch Changes
+
+- 37c3f2f: Udpated dependencies to support RC ima versions
+- 2e61a48: Built using new version of @ima/plugin-cli, js sources now include source maps
+- d08218b: Added type declarations generated from JSdoc
+- 6d47aaa: Side effects notation from package.json
+- 5717869: Added $dependencies typing comments for TS apps
+
 ## 1.2.3-rc.2
 
 ### Patch Changes